feat: Change SecretProbability type from decimal to int

+++ BlindTreasure.Domain/DTOs/BlindBoxDTOs/BlindBoxDetailDto.cs
@@ -13,7 +13,7 @@ public class BlindBoxDetailDto
     public DateTime ReleaseDate { get; set; }
     public BlindBoxStatus Status { get; set; } // ví dụ: Draft, PendingApproval, Approved...
     public bool HasSecretItem { get; set; }
-    public decimal SecretProbability { get; set; }
+    public int SecretProbability { get; set; }
     public bool IsDeleted { get; set; }
     public List<BlindBoxItemDto>? Items { get; set; }
 }
